Attribute "skip" is no initialized with value of property maven.test.skip
If one runs hibernate4:export with the switch "-Dmaven.test.skip=true", the
configured database was droped, although the execution of the tests was
skipped.

This might lead to serious data-loss, if one builds a project on the
production-machine. Hence, the excecution of hibernate4:export is aborted
now, if maven.test.skip is true.