Added the goal "update"
This new goal is not as useful, as it looks, because the quality of the
generated update-script heavily depends on the support for the JDBC metadata
API of the used driver.

As the added integration-test shows, this support is rather bad for the
H2-database, that is used in the test: a change to the length of one
column is ignored. Only the rename is reflected.

The same seems to hold for MySQL, as can be read here:
http://stackoverflow.com/a/12405147/247276